backend-scripts/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
appnope==0.1.4
APScheduler==4.0.0a5
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
babel==2.17.0
bardapi==0.1.23a0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
browser-cookie3==0.19.1
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chess==1.11.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.0.1
fastjsonschema==2.19.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozendict==2.4.1
frozenlist==1.4.1
fsspec==2025.3.0
gast==0.5.4
google==3.0.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.62.1
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
install==1.3.5
instaloader==4.11
ipykernel==6.29.4
ipython==8.23.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jellyfish==1.1.0
jieba3k==0.35.1
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
keras==3.2.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
loguru==0.7.2
lxml==5.2.1
lxml_html_clean==0.2.2
lz4==4.3.3
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
murmurhash==1.0.10
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.9.1
notebook==7.1.2
notebook_shim==0.2.4
nsepy==0.8
numpy==1.26.4
oauthlib==3.2.2
openai==1.59.8
opencv-python==4.11.0.86
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.25.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.6.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gemini-api==2.4.8
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2024.9.11
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
segtok==1.5.11
Send2Trash==1.8.3
setuptools==69.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.4.0
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==5.1.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.3.89
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
yake==0.4.8
yarl==1.9.4
yfinance==0.2.37
ytmusicapi==1.6.0
